requirements.txt
Django==3.2.6
django-filter==22.1
django-phonenumber-field==7.0.0
phonenumbers==8.13.0
django_debug_toolbar==3.8.1
djangorestframework==3.13.1
djoser==2.1.0
psycopg2-binary==2.9.5
djangorestframework-simplejwt==4.8.0